archivePackages {drat}	R Documentation

Move older copies of packages to an archive

Description

The function moves older versions of packages into a CRAN-style archive folder.

Usage

archivePackages(repopath = getOption("dratRepo", "~/git/drat"),
  type = "source", pkg)

Arguments

repopath
Character variable with the path to the repo; defaults to the value of the “dratRepo” option with “"~/git/drat"” as fallback

type
Character variable for the type of repository, so far “source”

pkg
Optional character variable specifying a package name(s), whose older versions should be archived. If missing (the default), archiving is performed on all packages.

Details

This function is still undergoing development and polish and may change in subsequent versions.

Author(s)

Thomas J. Leeper

Examples

## Not run: 
  archivePackages()   # archive all older package versions
  archivePackages(pkg = "drat")  # archive older copies of just one package

## End(Not run)

[Package drat version 0.1.0 Index]