packageHtml {drat}	R Documentation

Generate basic html website based on package metadata

Description

Generate basic html website with basic package info, optionally includes html manual. Package installation string can be included as well. CI environment variables for git branch and git commit will be included if available.

Usage

packageHtml(file = list.files(pattern = "*\\.tar\\.gz$"), repodir = ".",
  repo.url = character(), repo.cran = FALSE, manual = FALSE)

Arguments

file
An R package in source or binary format.

repodir
A local directory corresponding to the repository top-level directory.

repo.url
character, url to be used in repos argument in generated install.packages call for html website.

repo.cran
logical or character, default FALSE, if TRUE it will append repository list with CRAN repo in generated install.packages call.

manual
logical, default FALSE, when TRUE it will copy package html subdirectory to repodir and add a link to html/00Index.html. Additionally DESCRIPTION file will be copied so can be accessed from manual.

Value

TRUE is returned.

Author(s)

Jan Gorecki

Examples

## Not run: 
  packageHtml(file = "drat_0.1.0.tar.gz",
              repodir = "drat",
              repo.url = "https://eddelbuettel.github.io/drat",
              manual = TRUE)

## End(Not run)

[Package drat version 0.1.0 Index]